+++ delta-coverage-gradle/src/main/kotlin/io/github/surpsg/deltacoverage/gradle/DeltaCoveragePlugin.kt
@@ -6,6 +6,7 @@ import io.github.surpsg.deltacoverage.gradle.sources.lookup.KoverPluginSourcesLo
 import io.github.surpsg.deltacoverage.gradle.task.DeltaCoverageTask
 import io.github.surpsg.deltacoverage.gradle.task.DeltaCoverageTaskConfigurer
 import io.github.surpsg.deltacoverage.gradle.task.NativeGitDiffTask
+import io.github.surpsg.deltacoverage.gradle.utils.deltaCoverageConfig
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.Task
@@ -13,8 +14,11 @@ import org.gradle.api.plugins.JavaPlugin
 import org.gradle.api.tasks.TaskProvider
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import java.util.concurrent.ConcurrentHashMap
 
 open class DeltaCoveragePlugin : Plugin<Project> {
+
+    private val registeredViews = ConcurrentHashMap.newKeySet<String>()
 
     override fun apply(project: Project) = with(project) {
         extensions.create(
@@ -22,47 +26,56 @@ open class DeltaCoveragePlugin : Plugin<Project> {
             DeltaCoverageConfiguration::class.java,
             objects,
         )
+        CoverageEngineAutoApply().applyEngine(project)
+
+        val deltaTaskForViewConfigurer: (String) -> Unit = deltaTaskForViewConfigurer()
 
         extensions.configure<DeltaCoverageConfiguration>(DELTA_COVERAGE_REPORT_EXTENSION) { config ->
-            CoverageEngineAutoApply().apply(project, config)
+            // auto-register views from test tasks
+            registerReportViews(config, deltaTaskForViewConfigurer)
+        }
 
-            val nativeGitDiffTask = createNativeGitDiffTask(config)
-            val deltaCoverageLifecycleTask: Task = project.tasks.create(DELTA_COVERAGE_TASK)
+        afterEvaluate {
+            // Register custom views tasks
+            deltaCoverageConfig.reportViews.names.asSequence()
+                .filter { it != DeltaCoverageTaskConfigurer.AGGREGATED_REPORT_VIEW_NAME }
+                .filter { it !in registeredViews }
+                .forEach { viewName ->
+                    deltaTaskForViewConfigurer(viewName)
+                }
 
-            project.autoRegisterReportViews(config) { viewName ->
-                val deltaTask = createDeltaCoverageViewTask(viewName, config)
-
-                deltaCoverageLifecycleTask.dependsOn(deltaTask)
-                afterEvaluate {
-                    if (config.diffSource.git.useNativeGit.get()) {
-                        deltaTask.dependsOn(nativeGitDiffTask)
-                    }
+            deltaCoverageConfig.reportViews.named(DeltaCoverageTaskConfigurer.AGGREGATED_REPORT_VIEW_NAME) {
+                if (!it.enabled.isPresent) {
+                    it.enabled.set(registeredViews.size > 1)
                 }
             }
+            // Finally, register the aggregated view task
+            deltaTaskForViewConfigurer(DeltaCoverageTaskConfigurer.AGGREGATED_REPORT_VIEW_NAME)
         }
     }
 
-    private fun Project.autoRegisterReportViews(
+    private fun Project.registerReportViews(
         config: DeltaCoverageConfiguration,
         onView: (String) -> Unit,
-    ) {
-        val registerView = { viewName: String ->
-            config.reportViews.maybeCreate(viewName)
-            onView(viewName)
-        }
-        ViewLookup.lookup(this, registerView)
-        registerView(DeltaCoverageTaskConfigurer.AGGREGATED_REPORT_VIEW_NAME)
+    ) = ViewLookup.lookup(this) { viewName: String ->
+        config.reportViews.maybeCreate(viewName)
+        onView(viewName)
     }
 
-    private fun Project.createNativeGitDiffTask(
-        config: DeltaCoverageConfiguration,
-    ): TaskProvider<NativeGitDiffTask> {
-        return tasks.register(GIT_DIFF_TASK, NativeGitDiffTask::class.java) { gitDiffTask ->
-            val diffSource = config.diffSource
-            gitDiffTask.targetBranch.set(diffSource.git.diffBase)
+    private fun Project.deltaTaskForViewConfigurer(): (String) -> Unit {
+        val deltaCoverageLifecycleTask: Task = project.tasks.create(DELTA_COVERAGE_TASK)
+        val nativeGitDiffTask: TaskProvider<NativeGitDiffTask> = createNativeGitDiffTask()
 
-            diffSource.git.nativeGitDiffFile.set(gitDiffTask.diffFile)
-            gitDiffTask.dependsOn(JavaPlugin.CLASSES_TASK_NAME)
+        return { viewName: String ->
+            val config: DeltaCoverageConfiguration = deltaCoverageConfig
+            val deltaTask = createDeltaCoverageViewTask(viewName, config)
+            deltaCoverageLifecycleTask.dependsOn(deltaTask)
+            afterEvaluate {
+                if (config.diffSource.git.useNativeGit.get()) {
+                    deltaTask.dependsOn(nativeGitDiffTask)
+                }
+            }
+            registeredViews += viewName
         }
     }
 
@@ -73,6 +86,20 @@ open class DeltaCoveragePlugin : Plugin<Project> {
         val taskName: String = DELTA_COVERAGE_TASK + viewName.capitalize()
         return project.tasks.create(taskName, DeltaCoverageTask::class.java) { deltaCoverageTask ->
             DeltaCoverageTaskConfigurer.configure(viewName, config, deltaCoverageTask)
+            deltaCoverageTask.onlyIf {
+                config.reportViews.getByName(viewName).enabled.convention(true).get()
+            }
+        }
+    }
+
+    private fun Project.createNativeGitDiffTask(): TaskProvider<NativeGitDiffTask> {
+        return tasks.register(GIT_DIFF_TASK, NativeGitDiffTask::class.java) { gitDiffTask ->
+
+            val diffSource = deltaCoverageConfig.diffSource
+            gitDiffTask.targetBranch.set(diffSource.git.diffBase)
+
+            diffSource.git.nativeGitDiffFile.set(gitDiffTask.diffFile)
+            gitDiffTask.dependsOn(JavaPlugin.CLASSES_TASK_NAME)
         }
     }