+++ delta-coverage-gradle/src/testFixtures/kotlin/io/github/surpsg/deltacoverage/gradle/unittest/GradleUnitTests.kt
@@ -25,6 +25,7 @@ fun testJavaProject(
 
     it.customize()
 
+    it.evaluate()
     it.gradle.taskGraph.populate(FinalizedExecutionPlan.EMPTY)
 }